+++ tests/src/19-get-index-property.ts
@@ -2,21 +2,33 @@ import {
 	getType,
 	Type
 } from "tst-reflect";
-import * as ts from "typescript";
 
 test("getType<T>() gets index property", () => {
 	interface ObjectWithIndexProperty
 	{
 		a: string;
 		b: number;
-		[name: string]: string|number;
+		[name: string]: string | number;
+		[name: symbol]: boolean;
 	}
 
 	const type = getType<ObjectWithIndexProperty>();
+	
 	const properties = type.getProperties();
-	expect(properties).toHaveLength(3);
-	const indexProp = properties.find(prop => prop.name == ts.InternalSymbolName.Index)!;
-	expect(indexProp.type.types).toHaveLength(2);
-	expect(indexProp.type.types![0].fullName).toBe('String');
-	expect(indexProp.type.types![1].fullName).toBe('Number');
+	expect(properties).toHaveLength(2);
+	
+	const indexes = type.getIndexes();
+	expect(indexes).toHaveLength(2);
+	
+	const index1 = indexes.find(i => i.type.isUnion());
+	const index2 = indexes.find(i => i != index1);
+	
+	expect(index1).toBeTruthy();
+	expect(index2).toBeTruthy();
+
+	expect(index1!.keyType.isString()).toBeTruthy();
+	expect(index1!.type.isUnion()).toBeTruthy();
+	
+	expect(index2!.keyType.isSymbol()).toBeTruthy();
+	expect(index2!.type.isBoolean()).toBeTruthy();
 });